[refactor] TikkleListPage: class명에서 page빼기
