+++ Tikkle/Tikkle/TikkleListPage/TikkleListPage.storyboard
@@ -11,7 +11,7 @@
         <!--Tikkle-->
         <scene sceneID="s0d-6b-0kx">
             <objects>
-                <viewController storyboardIdentifier="TikkleListPageViewController" id="Y6W-OH-hqX" customClass="TikkleListPageViewController" customModule="Tikkle" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="TikkleListPageViewController" id="Y6W-OH-hqX" customClass="TikkleListViewController" customModule="Tikkle" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="5EZ-qb-Rvc">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
